requirements.txt
asgiref==3.10.0
autopep8==2.3.2
boto3==1.40.50
botocore==1.40.50
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.0
django-cors-headers==4.9.0
django-storages==1.14.6
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
mypy_extensions==1.1.0
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.5.0
psycopg==3.2.10
psycopg-binary==3.2.10
pycodestyle==2.14.0
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytokens==0.1.10
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
